algumas alteraçoes
