+++ src/index.html
@@ -13,13 +13,11 @@
   <script src="https://www.gstatic.com/firebasejs/7.23.0/firebase-auth.js"></script>
   <script src="https://www.gstatic.com/firebasejs/7.23.0/firebase-firestore.js"></script>
   <script src="https://www.gstatic.com/firebasejs/7.23.0/firebase-functions.js"></script>
-  <!-- <script src='/__/firebase/init.js?useEmulator=true'></script> -->
-  <!-- <script src="https://www.gstatic.com/firebasejs/8.2.1/firebase-app.js"></script> -->
 </head>
 
 <body>
   <div id="root"></div>
-  <script type="module" src="services/index.js"></script>
+  <script type="text/javascript" src="services/init.js"></script>
   <script type="module" src="router.js"></script>
 </body>
 </html>
