--- src/services/index.js
@@ -1,12 +1,0 @@
-// exporte suas funções
-
-export const firebaseConfig = {
-  apiKey: 'AIzaSyDNMVcpnbjVDYMArJM5XabYX5spXB8ANuw',
-  authDomain: 'rede-social-2c561.firebaseapp.com',
-  projectId: 'rede-social-2c561',
-  storageBucket: 'rede-social-2c561.appspot.com',
-  messagingSenderId: '528570691002',
-  appId: '1:528570691002:web:cbf56b96746afcaf0e0183',
-};
-  // Initialize Firebase
-firebase.initializeApp(firebaseConfig);
