+++ src/services/init.js
@@ -1,4 +1,4 @@
-const firebaseConfig = {
+export const firebaseConfig = {
     apiKey: 'AIzaSyDNMVcpnbjVDYMArJM5XabYX5spXB8ANuw',
     authDomain: 'rede-social-2c561.firebaseapp.com',
     databaseURL: 'https://rede-social-2c561-default-rtdb.firebaseio.com',
@@ -7,5 +7,5 @@ const firebaseConfig = {
     messagingSenderId: '528570691002',
     appId: '1:528570691002:web:cbf56b96746afcaf0e0183'
 };
-
-firebase.initializeApp(firebaseConfig);
+  // Initialize Firebase
+  firebase.initializeApp(firebaseConfig);
